+++ src/app/page.tsx
@@ -1,40 +1,12 @@
 "use client"
 
-import { useRef, useEffect } from 'react';
+import { MutableRefObject } from 'react';
+import { useInteractiveEffect } from '@/hooks/useInteractiveEffect';
 import Link from 'next/link';
 import '../../src/styles/gradient-bg.css';
 
 const Home: React.FC = () => {
-  const interactiveRef = useRef<HTMLDivElement | null>(null);
-
-  useEffect(() => {
-    let curX = 0;
-    let curY = 0;
-    let tgX = 0;
-    let tgY = 0;
-
-    function move() {
-      curX += (tgX - curX) / 20;
-      curY += (tgY - curY) / 20;
-
-      if (interactiveRef.current) {
-        interactiveRef.current.style.transform = `translate(${Math.round(curX)}px, ${Math.round(curY)}px)`;
-      }
-      requestAnimationFrame(move);
-    }
-
-    const handleMouseMove = (event: MouseEvent) => {
-      tgX = event.clientX;
-      tgY = event.clientY;
-    };
-
-    window.addEventListener('mousemove', handleMouseMove);
-    move();
-
-    return () => {
-      window.removeEventListener('mousemove', handleMouseMove);
-    };
-  }, []);
+  const interactiveRef: MutableRefObject<HTMLDivElement | null> = useInteractiveEffect();
 
   return (
     <div className=''>
