+++ src/hooks/useInteractiveEffect.ts
@@ -1,0 +1,36 @@
+import { useRef, useEffect, MutableRefObject } from 'react';
+
+export const useInteractiveEffect = (): MutableRefObject<HTMLDivElement | null> => {
+  const interactiveRef = useRef<HTMLDivElement | null>(null);
+
+  useEffect(() => {
+    let curX = 0;
+    let curY = 0;
+    let tgX = 0;
+    let tgY = 0;
+
+    function move() {
+      curX += (tgX - curX) / 20;
+      curY += (tgY - curY) / 20;
+
+      if (interactiveRef.current) {
+        interactiveRef.current.style.transform = `translate(${Math.round(curX)}px, ${Math.round(curY)}px)`;
+      }
+      requestAnimationFrame(move);
+    }
+
+    const handleMouseMove = (event: MouseEvent) => {
+      tgX = event.clientX;
+      tgY = event.clientY;
+    };
+
+    window.addEventListener('mousemove', handleMouseMove);
+    move();
+
+    return () => {
+      window.removeEventListener('mousemove', handleMouseMove);
+    };
+  }, []);
+
+  return interactiveRef;
+};
